rename etcd test to hacontrolplane
